+++ packages/stream_chat_flutter/lib/src/bottom_sheets/edit_message_sheet.dart
@@ -1,6 +1,37 @@
 import 'package:flutter/material.dart';
 import 'package:stream_chat_flutter/stream_chat_flutter.dart';
 
+/// {@template showEditMessageSheet}
+/// Displays an interactive modal bottom sheet to edit a message.
+/// {@endtemplate}
+Future<T?> showEditMessageSheet<T extends Object?>({
+  required BuildContext context,
+  required Message message,
+  required Channel channel,
+  EditMessageInputBuilder? editMessageInputBuilder,
+}) {
+  final messageInputTheme = StreamMessageInputTheme.of(context);
+
+  return showModalBottomSheet<T>(
+    context: context,
+    elevation: 2,
+    isScrollControlled: true,
+    clipBehavior: Clip.antiAlias,
+    backgroundColor: messageInputTheme.inputBackgroundColor,
+    shape: const RoundedRectangleBorder(
+      borderRadius: BorderRadius.only(
+        topLeft: Radius.circular(16),
+        topRight: Radius.circular(16),
+      ),
+    ),
+    builder: (context) => EditMessageSheet(
+      channel: channel,
+      message: message,
+      editMessageInputBuilder: editMessageInputBuilder,
+    ),
+  );
+}
+
 /// {@template editMessageSheet}
 /// Allows a user to edit the selected message.
 /// {@endtemplate}