+++ packages/stream_chat_flutter/lib/src/message_widget/message_card.dart
@@ -165,7 +165,7 @@ class _MessageCardState extends State<MessageCard> {
     return Container(
       constraints: const BoxConstraints().copyWith(maxWidth: widthLimit),
       margin: EdgeInsets.symmetric(
-        horizontal: (widget.isFailedState ? 15.0 : 0.0) +
+        horizontal: (widget.isFailedState ? 12.0 : 0.0) +
             (widget.showUserAvatar == DisplayWidget.gone ? 0 : 4.0),
       ),
       clipBehavior: Clip.hardEdge,