+++ packages/stream_chat_localizations/lib/src/stream_chat_localizations_fr.dart
@@ -625,4 +625,18 @@ Limite de pièces jointes dépassée : il n'est pas possible d'ajouter plus de $
   @override
   String get holdToRecordLabel =>
       'Maintenez pour enregistrer, relâchez pour envoyer';
+
+  @override
+  String get sendAnywayLabel => 'Envoyer quand même';
+
+  @override
+  String get moderatedMessageBlockedText =>
+      'Message bloqué par les politiques de modération';
+
+  @override
+  String get moderationReviewModalTitle => 'Êtes-vous sûr ?';
+
+  @override
+  String get moderationReviewModalDescription =>
+      '''Réfléchissez à la façon dont votre commentaire pourrait affecter les autres et assurez-vous de respecter nos directives communautaires.''';
 }